src/internal/cleanroom-internal/pyproject.toml
[project]
name = "cleanroom-internal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "hprabh", email = "hprabh@microsoft.com" }
]
requires-python = ">=3.10"
dependencies = [
    "requests==2.32.4",
    "opentelemetry-api==1.35.0",
    "opentelemetry-sdk==1.35.0",
    "opentelemetry-exporter-otlp-proto-common==1.35.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.35.0",
    "opentelemetry-instrumentation==0.56b0",
    "opentelemetry-instrumentation-fastapi==0.56b0",
    "opentelemetry-instrumentation-logging==0.56b0",
    "opentelemetry-instrumentation-requests==0.56b0",
    "opentelemetry-processor-baggage==0.56b0",
    "opentelemetry-proto==1.35.0",
    "opentelemetry-semantic-conventions==0.56b0",
    "opentelemetry-util-http==0.56b0"
]

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

